requirements.txt
torch>=2.4.0
Pillow>=10.2.0
numpy>=1.26.4
pytorch3d>=0.7.4
